Replace PtrElementHeap with completely non-generic heaps
Fixed error in PtrElementHeap and tried a few variants but results are not particularly interesting.

Introduced specialized heaps instead for the various element types.
Give a nice baseline for comparison and have (as expected) very good performance when
used in framework.

Switched to generating the element values using random() so we get the same values for every test.
Also, all element types now use the same values.